instancetypes: Move ENV constants into a more meaningful package
